git-svn-id: http://svn.codergrid.de/svn/sf/bruteripper/trunk@2124 7add9c9f-332d-0410-adcc-f7b1da7091d4

+++ src/org/esb/web/project/InputFilePanel.cpp
@@ -42,7 +42,7 @@ namespace org {
           std::vector<db::MediaFile>::iterator it = files.begin();
           //          if(files.size()==0)
           //            this->removeRows(0,rowCount());
-          if(rowCount()>files.size()){
+          if(static_cast<unsigned int>(rowCount())>files.size()){
             if(files.size()>0)
               removeRow(files.size()-1);
             else{
